requirements.txt
altair==5.1.1
APScheduler==3.10.1
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
dataclasses-json==0.5.8
dnspython==2.4.2
gitdb==4.0.10
GitPython==3.1.36
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-enum==1.5.1
mdurl==0.1.2
mypy-extensions==1.0.0
numpy==1.26.0
packaging==23.1
pandas==2.1.0

Pillow==9.5.0
protobuf==4.24.3
pyarrow==13.0.0
pycryptodome==3.18.0
pydeck==0.8.0
Pygments==2.16.1
pymongo==4.5.0
Pympler==1.0.1
python-dateutil==2.8.2
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
referencing==0.30.2
requests==2.31.0
rich==13.5.3
rpds-py==0.10.3
six==1.16.0
smmap==5.0.1
streamlit==1.26.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==4.3.1
urllib3==1.26.6
validators==0.22.0
watchdog==3.0.0
zipp==3.16.2
